Add a command line option for the test mode
In the test mode, EVS reference application assumes that current gear is
located in the reverse position and simulates corresponding behaviors.
This change implements a command line option to allow the users to set
this gear signal to configure the test mode operations.

To run the test mode with the park gear signal, the user can use below
command:
> adb shell evs_app --test --gear park

When this parameter is not given or set as reverse, EVS reference
application behaves the same as before.

Also, this change optimizes parameters in config.json for Osprey.

Fix: 163641719
Test: Run above examples
Change-Id: I4f801ad035083b4a016b7f9b4f80e2c09003d6d1