Fix a destruction of Enumerator objects
This change stops the statistics collection explicitly in the destructor
of Enumerator class, to avoid freeing StatsCollector object twice that
causes SIGABRT.

Also, this modifies the looper message handler to return an error if a
current event has been changed to terminate the collection while it was
in the queue.

And, this moves static Enumrator object creation to
LLVMFuzzerTestOneIput() and destroys a static Enumerator object
explicitly by setting a variable as null at the end of the function.

Fix: 162631113
Fix: 163813461
Test: Build evs_enumerator_fuzzer and run on the target device
Change-Id: I74b37dc6246575aab207ed1fcce3ef92762a6d9c