Remove race between "shutdown" and "cancel shutdown"
Most power requests come from the Vehicle HAL and if
more than one request is in the queue, we act on only
the most recent request.

After the VHAL requests SHUTDOWN_PREPARE, when we're done
preparing and we're ready to shut down, we queue a
WAIT_FOR_FINISH request (requesting shutdown). If the VHAL
decided to cancel the shutdown, their request might hit
the queue just before our WAIT_FOR_FINISH hits.

Because the "cancel" is older in the queue, it is ignored.

To fix this problem, we will not queue our WAIT_FOR_FINISH
if there is a "cancel" request in the queue already.

Fixes: 168853720
Bug: 164977558
Test: atest CarPowerManagementServiceTest#testShutdownCancel still passes
Test: A new test would be good, but difficult: request shutdown prepare;
      before that is ready to shut down, force CANCEL and WAIT_FOR_FINISH
      into the queue.
Change-Id: I4d4063fef9d93c73386eca0b61e83612fb09db76