Initial implementation of factory reset confirmation UI.
With device policy management support, it's possible that a factory
reset request is not originated by the user, and factory resetting
the device in this case could be a driving safety hazard.

So, rather than factory reset right way, CarService will show a
notification informing the user that if the device is not reset now,
it will be later.

Test: manual verification

Bug: 171603586

Change-Id: Ib046d168c3380bbe4a3923530860ba0acb8fb6ae