Add <uses-library> "org.apache.http.legacy" to Android.bp file.
The manifest for CarDeveloperOptions is merged from the manifest of its
static dependency Settings-core, which contains a <uses-library> tag for
"org.apache.http.legacy". Soong must know about this dependency in order
to pass verify_uses_libraries check and generate correct class loader
context for dexpreopt. Soong is unable to automatically deduce the
dependency from "libs" because it is not in "libs".

In the future, when b/176240706 is fixed, "uses_libs" entry for
"org.apache.http.legacy" should be removed.

Bug: 132357300
Bug: 176240706

Test: m out/soong/.intermediates/packages/services/Car/packages/CarDeveloperOptions/CarDeveloperOptions/android_common/manifest_check/AndroidManifest.xml
Change-Id: I07b8abafc10a878214c220cf945512e6960e4cbe