Switch CarDeveloperOptions from IConnectivityManager to VpnManager.
The VPN code is moving out of ConnectivityService to a new
VpnManagerService. Update Settings to call into the VpnManager
class instead of using the AIDL interface directly. This way,
the VPN code can be moved without touching Settings.

This patch is identical to aosp/1574947, which was merged into
Settings a few days ago, except some of the files were in
different directories. The patch was obtained by downloading the
diff from aosp/1574947 and applying the changes in each directory
in which the files were.

This patch contains no logic or behaviour changes.

Bug: 173331190
Test: m CarDeveloperOptions
Exempt-From-Owner-Approval: identical to Settings change
Exempt-From-Owner-Approval: code no longer exists in sc-dev or master
Change-Id: I19e937d75cfead91fe3822febaeb06142f042813
Merged-In: I4c4fbca57e3e52b5d20aee6144d3a1be3c5352c2