Establish a trivial CarTelemetryService.
It does not do anything and not wired into ICarImpl yet.
We will be adding to its implementation once we converge on design
of migrated Secular Trends a bit more. In addition migrates LogCat
related utilities and necessary protos.

Highlights of changes that had to be done out of necessity, because the
dependencies not being readily available in Android platform:
1) telemetry.proto is what used to be logprocessor.proto with important
change - I replaced timestamp field type with int64 to avoid dependency
on google3 proto.
2) LogcatReader.java. I replaced concurrent set with a regular one, but
made sure all access is done through methods that are all synchronized.
3) LogFilter.java. I replaced AutoValue with boiler plate code. Also
replaced ImmutableList with regular list + Collections.unmodifiableList.
4) Replaced references to LogProcessor with Car Telemetry service
instead.

Test: test migration will be done in the follow up CLs. This code is
changed only where necessary with equivalents available in the platform.

Bug: 180431013

Change-Id: If97889c7f1c7cf6364dfedf5ea95277b337fcb1c