Added auto start audio activity
Added audio activity that will automatically start playing audio on
startup. This includes managing its own audio focus request. Also added
different application that will start the audio application for the
current user assigned to the display. The latter first determines its
own display and uses that to determine the user assigned to the
display. It does this by using the occupant zone manager. Thus the
occupant zone manager should first be used to set up a profile user
for the desired display.

Test: make gcar_md_emu_x86
Test: create profile user and assign to display 1
Test: launch audio applicaton for user via: adb shell am start --user 13
	com.google.android.car.kitchensink
	/com.google.android.car.kitchensink.AudioAutoStartActivity
Test: launch occupant zone launcher via: adb shell am start --display 1
	com.google.android.car.kitchensink
	/com.google.android.car.kitchensink.OccupantZoneStartActivity

Bug: 179680322
Change-Id: If41b39870c2101176600af4b42fa972e780c1146