Re-add listener impl to CarService only when there are other listeners.
Manager's listener implementation is removed from CarService when there
are no more listeners for a specific resource type. This implementation
should be re-added to CarService only when there are listeners available
for other resource types.

This behavior is tested by verifying there are no more unexpected
interactions with CarService at the end of the unit tests.

Test: atest CarWatchdogManagerUnitTest
Fixes: 186865583
Change-Id: I3679973e7a8da29317a60b814289982621a9c084