Delay verifying watchdog daemon registration in unittest.
CarWatchdogServiceUnitTest verifies the registration with the daemon
asynchronously. This verification is flaky and fails when the acual
registration, that is done by a CarService's message handler, doesn't
complete within a timeout.

Post a message on the CarService's message handler and wait for the
message to execute before verifying the registration with the daemon.

Test: atest CarWatchdogServiceTest
Bug: 188059152
Fixes: 189233542

Change-Id: I1407ca6d11b30988a389db3b9fa8c86e45de82df