Create minimal ScriptExecutor java class to invoke script.
Amends IScriptExecutor invokeScript API to accept publishedData as a
bundle instead of byte[].
We do not parse the published input yet. It will be implemented
separately (see b/189241508).
Also adds necessary JNI layer.

Adds a unit test ScriptExecutorTest which starts ScriptExecutor
as an isolated process and binds to it.

Test cases push a simple Hello world script and test that it can be
loaded and run without exceptions. We don't test for output at this
point because it will be pushed asyncronously to CarTelemetryService
using IScriptExecutorListener callback. This functionality will be
implemented in the follow up CLs.

Bug: 187517413
Test: atest ScriptExecutorTest
Change-Id: Ida7c6ba9332532f3704a83a4ebfaf12dbec83bcb