Read the on-device XML configurations on initialization.
Update watchdog daemon to read the on-device resource overuse
configurations, that are stored as XML files, on initialization.

When latest configs are available under the data partition, watchdog
daemon uses the latest configs. Otherwise, the XML configurations from
the build are used. If all of these fail, watchdog daemon fallbacks to
using int max as the default thresholds.

When either of the vendor or third-party configurations aren't
available, watchdog daemon uses the system component's thresholds for
the vendor or third-party components.

Test: atest libwatchdog_test
Bug: 186445673

Change-Id: I94e08d618ff0cd164aadf4d37eae93e19b2dae67