Implement on_script_finished ScriptExecutor API.
This is a manual cherry-pick of the corresponding change
in sc-v2-dev with id: Ic1d85c82fc3aab5379cc2fc3b77402c7987f18d1

This API call is used for the user provided script to declare that
it has fulfilled its goal and finished the end of its lifetime and
the final results should be uploaded to the user.
The final results are returned using a Lua table which is then converted
to PersistableBundle by ScriptExecutor, before being returned to
CarTelemetryService via a callback.

Added relevant functional tests

Test: atest ScriptExecutorUnitTest:ScriptExecutorTest
Bug: 187517323
Bug: 193565932
Change-Id: Iec0d98a5469b4f3a3af334b4f7e537b6050fba42