Fix broken ScriptExecutor tests.
During manual cherry-pick of changes from master to sc-v2-dev
some new deltas were created because ScriptExecutorTest was failing
unexpectedly. Compared to master, we also needed to change how we bind
to ScriptExecutor from ScriptExecutorTest. In particular, it had to be
bound as a system user.

Currently, ScriptExecutorTest is failing in master. Binding as a system
user fixes the problem and the test has been verified to pass.

Test: atest ScriptExecutorUnitTest:ScriptExecutorTest
Bug: 198477176
Change-Id: I5e7e1bdcada2f32e8e5e329e7fce0bbba69978c3