Overlay config for foreground DA components
This enables the notification system bar button to toggle the foreground DA. Namely, if the notification center is open, clicking the notification icon again will close the panel. (Maintains behavior that tapping notification icon if notification center is not open will open the foreground DA if not open and launch the notification center).

Bug: 197366554
Test: manual
Change-Id: I2ba9caa66d412633278eea7804a37eb74e24bb0d