Fixed call and ringer focus request issue
Fixed call and ringer focus request issues where the telephony stack
uses the same focus request listener for the ringer and call. If the
ringer focus request is still active car audio service rejects the
focus request given that the client id are the same for both
requests. The fix, verifies if the client id's are the same as the
telephony stack and replaces the current request witht the incoming
one. The latter occurs only if both requests are either ringer or call
contexts, otherwise the request is rejected. Also refactored
AudioServiceHelper to AudioManagerHelper.

Bug: 182547473
Test: - Pair bluetooth phone to car
  - Make a call from an external phone
  - Accept call either on car UI or phone
  - Verify no error occurs in call or logcat, the following line shoud
    not appear in the logcat:
  	E CarAudioFocus: Client AudioFocus_For_Phone_Ring_And_Calls
	has already requested focus
Change-Id: I78f8e4b4067ab955554dadd1bcd96718215969b8