Add new status UPLOADED_BEFORE
This case can happen when a bugreport was uploaded before,
but somehow the app failed to update its status to
UPLOAD_SUCCESS. And the app tries to upload it again,
but back-end (GCS) doesn't allow rewriting the file.

This status will help when this case happens and bugreport
may not actually have been uploaded. It temporarily persists
zip files for investigation.

Context: ag/16272708

Bug: 203708797
Test: atest BugReportAppTest
Test: manually teted on a device
Change-Id: Id4894ad3f0fad4f099991974ad3d85a8a50fef68