Use a new Parcel for largeParcelable.
Use a new Parcel for largeParcelable serialization so that we would
not try to call marshal function on the output parcel which might
contain data that could not be marshalled, e.g. binder interface.

This CL also updates the serialization function from writing to
a parcel to returning a shared memory file descriptor.
I want to avoid directly manipulating parcel that would
be sent across binder interface since it is hard to make the parcel
compatible with the system implementation. By updating the API, the
function would still follows the default system way of turning
parcelable into a parcel while sending across binder.

This would add an additional copy of small payload, but since the
payload is small, the additional cost is negligible compared to
other costs. I see no significant difference in performance test.

Test: AndroidCarApiTest:com.android.car.internal.test
Bug: 209506058
Change-Id: Ie8ff92df13cbd6d085737183b86f689e093aec59