Replace primary key in UserPackageSettings table.
- Add an autoincrementing integer as the new primary key for
UserPackageSettings table and make dependent table use this primary key
as their foreign key.
   - IoUsageStats table used the rowid of UserPackageSettings table as
   one of its primary keys and had a foreign key dependency on
   the rowid. But using rowid as a primary key is discouraged because
   the rowid of an entry is not guaranteed to be the same and can
   change when the table shrinks.
- Upgrade the database version and implement the onUpgrade method to
migrate the settings from the old UserPackageSettings table to the
new UserPackageSettings table.
  - IoUsageStats table has mappable entries only for the current date
  because the rowids of the historic stats are not mappable. So, drop
  this table on upgrade.
- Update the logic to save entries in UserPackageSettings table such
that any update to existing entries are done via DB update SQL and new
entries are added via DB replace SQL. This is to ensure the primary key
doesn't autoincrement for existing entries. Otherwise, dependent entries
in other tables will be deleted.
- Enable foreign key constrain in the database, so IoUsageStats entries
are in sync with the UserPackageSettings entries.

Test: atest WatchdogStorageUnitTest
Bug: 213190011

Change-Id: I34a3aba2ff337abcbdc9cde049c7829428391344