Make LargeParcelableBase compatible with stable AIDL.
This CL makes LargeParcelableBase and its subclasses generates a
parcel that is compatible with the Stable AIDL. When serializing
a stable AIDL, we store the parcel that serialized the whole
parcelable, including data size, payload and a file descriptor
into the shared memory file. We need to do the same for
LargeParcelableBase. Prior to this CL, LargeParcelableBase only
serializes the payload to the memory file.
This CL also fixes a bug where a fd is double-owned.

Test: atest AndroidCarApiTest:com.android.car.internal.test
Bug: 210063973
Change-Id: Ie674c550b701b415149bb4b6c788b50e7777d90f