Implement MemoryPublisher and handle publisher failure
This CL implements MemoryPublisher and adds kitchensink tests.

MemoryPublisher proto contains the snapshot frequency and the max
snapshot count. Only one subscriber for MemoryPublisher is allowed.
Otherwise an exception is thrown.

Previously the exception is silently ignored in DataBroker, this CL
makes CarTelemetryService handle the exception in addMetricsConfig.

This CL allows publishers to end a MetricsConfig once a terminating
condition has been met (the terminating condition is defined in the
MetricsConfig itself) and handles publisher failures by constructing
an error report and terminates the MetricsConfig.

This CL refactors some code by removing DataBrokerController and moving
its responsibility to CarTelemetryService.

Bug: 229134432
Fixes: 193680465
Fixes: 191378559
Test: m -j
Test: atest com.android.car.telemetry
Test: atest CarServiceTest:CarTelemetryManagerTest
Test: atest CtsCarTestCases:CarTelemetryManagerTest
Test: atest CarTelemetryHostTest
Test: kitchensink

Change-Id: Ia27806be2bd77b4b652dda49f33b1e79f2612dee