Refactor Audio Focus Exclusive Interaction Tests
The tests in CarAudioFocusTest about exclusive
interaction are moved to
CarZonesAudioFocusExclusiveTest. The tests are
parameterized with the usage, client id and client
uid of the two requests tested. The first request
should get audio focus, and then be excluded by
the second request. After the second request is
abandoned, the first request should get the focus
again if the gain is transient.

Bug: 219109875
Test: m -j, atest com.android.car.audio
Change-Id: Iea0867b9e5fe8015dd30214025e487d80c02ccc0