Replace 300+ line switch with automated int value to String system.
Currently the VehiclePropertyIds test has a 300+ line switch statement which requires manual maintenance and synchronisation of values whenever a new identifier is added. This CL removes the requirement for manual maintenance.

This CL puts in place a system which uses the field declarations in the class to provide a String representation of an int property which removes the need for a manually maintained switch statement. Using this method reduces the chances of manual errors causing discrepencies between the field declarations and the results of the switch statement (currently only a subset of return values are checked in the test suites). It also removes the need to manually add values to tests because, due to the method employed in this CL, it would not be possible for an Id declared as a public static final to be translated incorrectly.

Using lazy initialisation allows us to take advantage of the fact this method isn't called in many places (https://cs.android.com/android/platform/superproject/+/master:packages/services/Car/car-lib/src/android/car/VehiclePropertyIds.java;l=30;bpv=1;bpt=1 indicates it's only used in tests and when dumping values). This means that for almost all production usage there would be a closely comparible memory usage because the SparseArray would not be populated.

While this approach would be slower for the first call due to the need to populate the SparseArray, the subsequent performance would be similar due to the binary search used in both the `sparse-switch` (created by a switch statement of this size) and the SparseArray (used in this CL) using a binary search for map the proprty value onto either the code (for sparse-switch), or the return value (for the SparseArray).

Overall this CLs benefits for readability, and protection against the potential for manually introduced bugs, seems worthwhile to me.

Test: atest AndroidCarApiTest
Fixes: 234132292

Change-Id: Id23347fc35c1160025b95c63f04d1197fa365a90
(cherry picked from commit 8e6e24f22c5e3c7da4cfab9ef4b8c571f305c127)