Clarify the string for "GCS Upload pending"
From discussion on the linked bug, this state means that we are either currently uploading, or waiting for network before uploading.

This is a somewhat messy state, but it's better to have messy text accurately reflect that messy reality, rather than make it sound like the upload is waiting, when the upload is happening now.

Also, update some comments, to indicate they are one among many of the terminal states, not the (singular) final state.

BUG=233560534

Change-Id: I55d8b832520e0ad67ae1883675f0425e0ff5d83c