Move title bar to own Z layer
In certain applications, the touch area of the application may attempt
to extend into the title bar area since they are on the same layer. This
can prevent title bar touch events from being registered. By moving the
title bar to it's own Z layer above the foreground layer, this can be
avoided.

Bug: 238359700
Test: manual
Change-Id: I98b937e79da982861bde06d1054aba41e5acdf11