Implement public cartelemetryd AIDL interface for filtering data
This CL implements cartelemetryd's public AIDL interface for adding a
ICarTelemetryCallback.

When CarTelemetryService adds or removes CarData IDs, the callback will
be notified based on the IDs that client is interested in.

Clients can use this information to decide whether to write the data to
cartelemetryd, which allows data filtering at the data source.

The sample client is also updated to add and remove callback.

Bug: 230907689
Test: m -j
Test: atest cartelemetryd_impl_test
Test: manual test with sample client

Change-Id: I6d59ce43ded0a04731c152bc79f9e849c7b5bdd3