Added new annotation and enums for car / platform versions .
This change will:
- Make the annotations cleaner.
- Make it easier to mark methods that require a pre-release API.

For example, instead of

  @AddedIn(majorVersion = 33, minorVersion = 1)
  @MinimumPlatformSdkVersion(majorVersion = 33, minorVersion = 1)

A method would be annotated with:

  @ApiRequirements(minCarVersion = CarVersion.TIRAMISU_1,
                   minPlatformVersion = PlatformVersion.TIRAMISU_1)

Notice the use of the min prefix - it could allows us to add a max in
the future (in case an app is planned to be removed).

Finally, this change also moved the existing constants into
ApiVersion.VERSION_CODES (to be consistent with Build.VERSION_CODES).

Test: atest android.car.cts.ApiVersionTest android.car.cts.CarApiVersionTest android.car.cts.PlatformApiVersionTest android.car.AnnotationTest
Test: m -j android.car-module.stubs.source-update-current-api android.car-module.stubs.source.system-update-current-api android.car-stubs-docs-update-current-api android.car-system-stubs-docs-update-current-api android.car-test-stubs-docs-update-current-api android.car-module.stubs.source.test-update-current-api

Bug: 240471760
Bug: 240343308

(cherry picked from commit 465d12af951d5f6d99f1166b3eb02db06c4ca54d)
Change-Id: I1a4184a96314eae0503eb1f4d59e40e254d8f666