Expand sample client for cartelemetryd to send multiple batches of data.
Includes changes to a sample cartelemetryd client that allow to send a
batch of data with arbitrary payload size, frequency and batch size.
Modifies shell commands for CarTelemetryService to be able to read data
that comes in a batch, and measure how long it takes to read such a
batch.

Adds additional logging to carlemetryd when it fails to push cardata to
CarTelemetrydPublisher.

Test: Run sample client according to packages/services/Car/cpp/telemetry/cartelemetryd/sampleclient/README.md
Bug: 233607449

Change-Id: I7d3a6529ecc4b869558c07ee30801b0b1330b502