Improved tool to generate test classes file and hidden APIs
The tool will help to keep the class list for annotation test updated. Eventually we would merge this tool to metalava if possible.

The tool also generate the complete list of APIs (including hidden) for car-lib and car-built-in. This file can be used for
 - To check if there is any new hidden API added
 - To check if newly added API are annotated with correct car version.

To run tool:
 - build: m -j GenericCarApiBuilder
 - run: GenericCarApiBuilder <option>

Supported options are:
  "--print-classes-only";
  "--update-classes-for-test";
  "--generate-full-api-list";

Bug: 243164930
Test: atest android.car.AnnotationTest
Change-Id: Ic6cedecac52b2224c68e452677765479907d200e