Added supplementary annotations to ApiCheckerRule.
By default the expected behavior of a test (under that rule) is:
- if the device is supported, test should run and pass
- if the device is not supported, test should throw
  PlatformMismatchException

There are scenarios where these behaviors don't apply. For example,
if the test is exercising a callback that's only sent on supported
versions, it would not throw an exception on unsupported versions.
So, in these cases, the test must be split in 2, one annotated with
@SupportedVersionTest and another with @UnSupportedVersionTest.

Example:

@Test
@SupportedVersionTest
public void testCallbackIsSentOnSupportedVersions()

@Test
@UnsupportedVersionTest(behavior=EXPECT_PASS, supportedVersionTest="testCallbackIsSentOnSupportedVersions")
public void testCallbackIsNotSentOnUnsupportedVersions()

Test: atest android.car.test.ApiCheckerRuleTest
Bug: 242315785

Change-Id: Ie60bef7c96730dcec83199c83b7d7c271f36a59b