Launch simple user picker in CarUserService and move user-zone assignment to when the user is starting.
simple user picker will be launched when a display does not have
user allocation after the system boots to allow a selected passenger
to start on that display.

The user assignment to zone is moved to when the visible user is
starting so that only startUserInBackgroundOnSecondaryDisplay needs
to be called for visibility assignment and assignVisibleUserToOccupantZone
will happen under the hood. Updated user picker accordingly and added
a button to refresh the user-zone assignment.

This change also includes the following fixes:
- Makes sure that users started via. the user picker will have their
  launcher launched.
- The user picker can show up on multiple displays simultaneously.
- "Trampoline pattern" to relaunch the user picker as user 0 for it to
  stay after stopping user.

Bug: 244370727
Bug: 253264316
Test: Boot the system and see user picker showing up on display with no
user allocation.
Test: Stop a passenger user and see user picker stay for switching user.
Test: Perform visibility assignment in user picker and see launcher
launched for the assigned users.
Test: atest CarUserServiceTest
Test: atest UserAssignmentTest

Change-Id: Iac971f990d5b14d0e24047d50b32347c9975678c