Fix resource leak in VehicleStubTest.
Previously we have a bug where the handler thread is waiting on
a countdown latch which is counted down by the same thread. Because
the await is posted first, and is blocking the countdown, it will
hang forever. This causes the thread to always hanging. Because our
test case does not wait for all thread to finish, so our test finishes
successfully with the thread still running.

This will then cause the next test that extends
AbstractExtendedMockitoTestCase (testShowDisclaimerWhenIntentReceived)
to hang because while it is ending, it will call
completeAllHandlerThreadTasks which will wait for
all live threads to finish and will hang forever.

Test: atest CarServiceUnitTest
Bug: 250650721
Change-Id: Ie19dc37d761b756ac1515eb34c8d73d182d7a220