Fix focus jumping to Location Data View when changing day/night lights
When the NIGHT_MODE Vehicle Property changes, the system UI theme is updated accordingly. In cars this can be done by adjusting the light knob to on or off. By changing this value it causes the KitchenSink Activity to restart thus causing a focus jump despite there not being dedicated light and dark modes in KitchenSink. This change makes it so that the app handles the configuration change and does not restart.

Bug: 238763138
Test: manual test
Change-Id: Idf242418f8d2396c641efbcd4e6ed9be34d04c29