Added preconditions checks to fade and balance and four unit tests to car audio service
Added precondition checks for fade and balance for the ranges of -1 to
1 and changed the corresponding carAudioManager documentations
associated with the changes. Added set balance toward right and set
fade toward front to car audio service for test values of 0 and
verifies that the control wrapper's set fade toward front and set
balance twaord right gets called.

Bug: 255446023
Test: atest CarAudioServiceUnitTest, verified that method coverage is
called for those methods in local code coverage

Change-Id: I464a5d7bdb65b53a42c316a9c071abaab00548ed