Made car audio context class non static
Refactored car audio context class into non static. It now takes a
list of car audio context info. That list is used as the grouping of
audio attributes into logical sets which define the car audio contexts.
In later updates this will be allowed to be read from the car audio
configurationfile, which can be defined differently for each target.

Bug: 248106031
Test: atest com.android.car.audio
Test: atest AtsAudioDeviceTest CarAudioManagerSystemTest
Test: atest CarAudioManagerTest CarAudioFocusTest
Change-Id: Ifed569334ff061229e7c8cdeb7153900db2c6883