Add check for supported vehicle property ID.
Since car service is mainlined, we will update car service
outside android release and this will cause car app, car
service and VHAL to based on different property version.

This CL adds two checks. One is to check all the request
from car application to car service that the property ID
is supported by the car service. This is to prevent a newer
version of car app to use unsupported property in a lower
version of car service.

The other is to filter the properties returned from VHAL and
only expose properties supported by car service. This is for
situation were VHAL is using a newer version of VHAL interface
but car serivce is older.

In all cases, car service will only show support for properties
that are defined in its VehiclePropertyIds list to car app.

Bug: 255415147
Test: atest CarPropertyHelperUnitTest
atest CarPropertyManagerTest
atest CarPropertyManagerUnitTest

Change-Id: I85f80b5dc60c1fea0daa79f3f75fcec0160d4eca