Update isProfileSupported to check both devices supported features
We use the local set of UUIDs that our device claims to support to make
sure what we're about to connect is co-supported between the two
devices.

Previously we use CDD requirements to more or less assume that certain
profiles will be supported as a car. However, experiments are being done
with optional profiles and this pattern will help enable that going
forward.

Bug: 260726996
Test: atest CarServiceUnitTest

Change-Id: I1652592486e06bef230051f5b5b79b12a243a8c2