Introduce BinderKeyValueContainer
BinderKeyValueContainer is a map-like container to hold client's
binder interface. The item in the container will be removed
automatically once the associated binder is unlinked (dies).

This container is mainly used by CarOccupantConnectionService, which
needs to hold the callbacks from the client.

This container is also used by AbstractReceiverService. AbstractReceiverService runs on the receiver app side and needs to
hold the callbacks of the receiver endpoints. Although it is unusual, the process that registered the callback might be different from
the process that AbstractReceiverService runs in. If that process is
dead, when AbstractReceiverService invokes the dead callback, a
DeadObject exception will be thrown. So AbstractReceiverService needs to use this container.

Bug: 257117236
Test: atest com.android.car.internal.util.BinderKeyValueContainerTest

Change-Id: Ic8c4672c8404f523af04a65907962c5132de81d6