Fix NoSuchMethodError in CarMediaService
UserManagerHelper#getMaxSupportedUsers is added in U and can cause
NoSuchMethodError if the platform version is older than U.
We could add version checking to avoid this error, but it will also need
a fallback logic for older versions. It is not worth the complicated
logic because here it is only used for the initial capacity of a sparse
array that can be dynamically resized.
Therefore simply replace it with UserManagerHelper#getMaxRunningUsers,
which is good enough for this case.

Fixes: 266946454

Test: atest CarMediaServiceTest

Change-Id: Idf1a654afe5919386bf4b1f4dd13d303c288c01a