Post health checking to front of CarService main handler queue.
CarService should respond immediately when the CarWatchdog daemon pings
for health checking. If the main thread's handler queue (for current
time) is filled up with multiple messages, the health checking request
will be delayed until all prior messages are processed.

Avoid this by posting the health checking message to the front of the
CarService's main handler queue. This will allow CarService to respond
as soon as the current message is processed.

Test: atest CarWatchdogServiceUnitTest
Bug: 266965586
Bug: 262192888
Change-Id: If42f670ab061771ef1cf64e4d79c940e5285c1ca