Implement getInitValue for setAsync.
Implement the logic to get the current value before sending the
setAsync request. If the current value is the same as the target
value, there won't be a property update event. If we get the
initial value okay and it is the same as target value. Also if we
see the set value result is success, we will call the success
callback to client.

Test: atest PropertyHalServiceTest
Bug: 264719384
Change-Id: If055f82d8c10aad47bee53e6575e9d1e636f97db