Implement setAsync prop subscription.
Add internal subscription to property for async set property
operation. The internal subscription will track whether the property
is updated to the target value. If the property is updated to the
target value and the set operation succeeded, we will call the
success callback for the async set.

This internal subscription must not affect the external subscription
issued by the client, so we use a separate map to track the internal
subscriptions.

Test: atest PropertyHalServiceTest VehicleHalTest
Bug: 264719384

Change-Id: I565347a96180a5673336598eb0eade506f14b932
(cherry picked from commit 84343236db7c57c3e9d69badc02278c6d37021f9)