Limit sync operation to be 16 at-a-time.
If we do not limit sync operation count. They might use up all the
binder threads and all be blocked, while there is no thread left
for callback to arrive to unblock them.

The number 16 is just arbitrally picked. It just need to be less than
31 so not all threads are blocked.

Test: atest CarPropertyServiceUnitTest
CTS CarPropertyManagerTest using 32 thread pool in the client.
Bug: 268533181

Change-Id: I5fa521097b3fd28139c64d59e4a5d9ca6cebc6d3