Added audio zones mirror management
Added audio zones mirror management:
- Audio mirror status callback management
- Audio routing and partial focus management
The focus management currently assumes we are using a single zone
as focus source. Need to change to manage focus in multiple zones.
Also added logic to reject mirroring:
- If primary audio share enable, primary audio zone share must be
  first disabled.
- If audio mirror is set:
- - If current mirroring config is the same, the request is approved
    with no further changes.
- - If mirroring config is different, the new config will be allowed
    only if the config is expanded (addeing more zones).
- - If the mirroring config attempts to change a different config
    the change will be rejected. This means that merging two
    configurations is not allowed, must exit the current mirror before
    mirroring in other zones is allowed.
For the latter, the enableMirrorForAudioZones API only allows adding
more zones to the audio mirror configuration, can not remove. The
disableAudioMirrorForZone API will be allowed to remove a single
zone. We will also expand the APIS with disableAudioMirrorForZones
to allow for disabling multiple zones.

Bug: 265973263
Test: atest com.android.car.audio
Change-Id: I95ad9da016bca3c03a1b4c27151c9701b229f64a