Remove redundant registerConnectionStateCallback() and unregisterConnectionStateCallback()
These two methods provide an asynchronous way to check whether it is
connected to a given occupant zone. We introduced them because we
thought the client app might need to check that before sending payload.
But we already got isConnected() to check that synchronously.
To simplify the communication API, we can just remove them. If we find
that app developers really need them, we can add them back in the next
release.

Bug: 257118072
Test: manual
Change-Id: I32750e9d0d80bf3898b1128935aeb36bd7c1701c