Remove the call to connect to the client
The related EATT/CoC issues have been fixed and the subsequent ACL
disconnection have been root caused. This line of code is not required
anymore.

As well, _keeping_ the line of code can cause authentication requests to
go out from the provider to the seeker (when the provider has a link key
and the seeker does not). These can go out over BR/EDR as well, because
TRANSPORT_AUTO is used (no API exists not to) and the BT stack may
direct our connection out BR/EDR. Altogether, this can conflict with the
Fast Pairing process that's being  kicked off by the seeker.

Bug: 269670419
Test: Fast Pair with a seeker device when the provider has a link key
Change-Id: I2fa5bab300c6fd08182584baa2508799410454c6