Cleanup home launch for visible user in CarOccupantZoneService
Previously startHomeForUserAndDisplay is called in multiple places which
is redundant:
CarUserService#onUserVisible -> assignVisibleUserToZone ->
CarOccupantZoneService#assignVisibleUserToOccupantZone ->
startHomeForUserAndDisplay

CarUserService#onUserUnlocked -> startLauncherForVisibleUser ->
startHomeForUserAndDisplay

Now we clean up the code path in CarOccupantZoneService and only leave
the one in CarUserService on user unlocked. Note: this effort will also
facilitate work in passenger only system as all occurrences of
startHomeForUserAndDisplay need to be fixed for visible user on the
default display.

Bug: 243967195
Bug: 266473227
Test: atest CarOccupantZoneServiceTest
Test: atest CarOccupantZoneManagerTest
Test: manual on the MD emulator
Change-Id: Iba7d368849d93d68c06d0091a1a573e772f504fe