Fix unregister of volume callback when exception
The current logic removes the client callback from the list
and if the list is empty unregisters aidl callback  with the
car audio service. However, if the call to service results
in an exception, the aidl callback is retained.
Since the callback list of the car audio manager is empty,
the next call from client to register callback will result
in a new aidl callback registration with car audio service.
The orphaned callback is left in the car audio service
forever (or until the client dies).

Reordered the unregistration of volume callback to handle
such cases. If the request results in an exception, the same
is thrown to the client without removing the callback. The
app can retry to remove the callback.

Bug: 261647905
Test: atest com.android.car.audio

Change-Id: Iedd848ca2279588ddd851c78ac3c0b31957bf028
Signed-off-by: Raj Goparaju <rajgoparaju@google.com>