Fixed broken links in all android.car files.
All links in android.car are fixed except for some links that lead to
nonexistent/missing code that is not owned by the VAC team. These were
left as they were because the owners of the link are the ones best
equipped to change the link to its appropriate location.

Bug: 266147450
Test: manual build
Change-Id: I0409f6f257b436cb597e15839f033a01386bb928