Implement registerStateCallback() and unregisterStateCallback()
This CL also introduced CarRemoteDeviceService to implement the
API defined in CarRemoteDeviceManager because:
1. It makes more sense to have a separate car service for a manager
2. The implementation of CarRemoteDeviceManager is irrelevant to the
implementation of CarOccupantConnectionManager, so having them in
two separate services makes the code easier to read and maintain.

As a side effect, some methods are migrated from
CarOccupantConnectionService to CarRemoteDeviceService, and some
tests are migrated from CarOccupantConnectionServiceTest to
CarRemoteDeviceServiceTest.

Bug: 257117236
Test: atest com.android.car.CarRemoteDeviceServiceTest
Test: atest com.android.car.occupantconnection.CarOccupantConnectionServiceTest
Test: atest com.android.car.CarRemoteDeviceManagerUnitTest
Test: atest com.android.car.SystemActivityMonitoringServiceUnitTest

Change-Id: I3ad909a7fa87703692620617181dfe122bdcf9eb