Implement onDisconnected() and onFailed() correctly
onDisconnected() should be called when an established connection
is terminated, while onFailed() should be called when there is
an error during establishing connection.

When the sender requests a connection, if the receiver zone is not ready
for connection, car service should call onFailed() with
CONNECTION_ERROR_NOT_READY instead of throwing an IllegalStateException.

Fixes: 277238124
Bug: 257117236
Test: atest com.android.car.occupantconnection.CarOccupantConnectionServiceTest

Change-Id: I23fcfdbd71c7dc11bf2b812fa9068b66395c3034