Fix NullPointerException on CarVolumeGroupMuting
When group muting is disabled, vendor audio control hal should
not initiate callbacks with reasons that results in group mute.

However, IAudioControl interace does not prohibit this happening
and there is no discovery available for vendor audio control hal
to check if group mute is disabled.

Therefore, it is prudent that car audio service handle such a
case gracefully:
- Hal mute state is only updated if group muting is enabled
- Handle group muting only if group muting is enabled

Bug: 277296021
Test: atest -c com.android.car.audio

Change-Id: I8a3135415dba18adb2647699d60fd6440857f5e0