Redefine CPMS proto to reduce string parsing.
The previous CPMS proto definition included string fields that would
still require additional parsing upon reading them in order to extract
information, which is one of the main things that was trying to be
avoided by dumping to proto instead of text. By restructuring some of
the proto definitions, especially regarded to power policies, the need
to process strings upon reading them should be lessened.

Bug: b/281730477
Test: builds okay
Change-Id: I46ae06163137f4cdb5fb0a21ed9d1fc19cef3882