Handle a display state change once
CarEvsService sometimes receives duplicated notifications of a display
state change. When this happens, CarEvsService creates multiple
instances of the camera preview activity but closes only the instance
that created last.

To avoid this scenario, CarEvsService stores the latest display state
change it processed and ignores notifications with the same state.

Bug: 285050051
Test: Run a test sequence described in a bug.
Change-Id: Id13872f67856817872ad2e2de541634d5e1b3b9c