Update retry strategy for tryToRebind()
Previously the retry strategy was retrying after 5 seconds up to 5
times, which was not enough for services that require longer time, such
as VMS service that depends on CMSCore update.
Instead we are adopting exponential backoff, with the initial
delay of 4 seconds, doubling the delay after each failure, up to 6
times, which covers 4+8+16+32+64+128 = 252 seconds.
The max retry number is configurable.

Bug: 288064318

Test: atest VendorServiceInfoTest VendorServiceControllerTest
      Manually update GMSCore and check the logcat

Change-Id: Iaad6317493aa48722cb9ab3c98316523a4923674