Add PairSparseArray class
A SparseArray mapping two integer keys to an Object. It encodes the
32 bit integer keys into one 64 bit long with the first key in the most
significant 32 bits and the second key in the least significant 32
bits. Underneath, it uses a LongSparseArray to map the encoded long to
the Object. PairSparseArray handles encoding and decoding the keys.

The main use case for this in CarService is that we have many data
structures mapping property and area IDs to objects. Currently, we
either use nested SparseArrays or an ArrayMap with Pair<Integer,
Integer> as the key. The advantage of PairSparseArray is that it can
store both the property and area ID without autoboxing and without
multiple layers of maps.

Bug: 294574166
Test: atest PairSparseArrayUnitTest
Change-Id: I8a432cfbbd7b6740e0837ad1782acb47a260cd8d