Optimize add pending request logic.
When adding pending request to the pool, directly add all the
requests instead of adding one by one. This fixes a flaky test
where there is a chance when the first request is added and time
out message is triggered before the second request is added, causing
two callback to be invoked instead of one.

Test: atest -s emulator-5554 AidlVehicleStubUnitTest#testGetAsyncAidlTimeout --iteration 200
Bug: 297590027
Change-Id: I6ecc20e27b2a024ed3ec05c7be6b5ced9d1c5e3f