Fix race condition in syncLooper in WatchdogProcessServiceTest
WatchdogProcessServiceTest#syncLooper had a race condition where a
message was sent to a message handler and then immediately it waited for
the mutex lock to be released once the handler handled the sent message.
The issue was that it was possible for the handler to handle the message
and call notify_all() on the lock before the lock wait_for() method is
called. This resulted in the lock waiting for the timeout before being
released.

Test: atest WatchdogProcessServiceTest --rerun-until-failure 200
Bug: 299676049
Change-Id: I35db20d28e925f559fc4a1492e47d8094d4b3bb7