Throw exception from VehicleHal subscribe.
Previously VehicleHal subscribe/unsubscribe property will catch
all exceptions from VHAL, causing the higher layers to assume
the subscription succeeded even though it may failed. This caused
higher layer to always update their internal state and causes
inconsistency between the internal state in car service and the state
in actual VHAL. This causes a bigger issue where even if the user
retries the operation, higher layer will assume the operation already
succeeded and do not send the retry request to VHAL.

This CL updates VehicleHal to throw the exception and updates the
higher layer. For each layer, if the operation fails, it should resume
to its previous state so that the higher layer may retry.

Test: atest VehicleHalTest PropertyHalServiceTest
Bug: 301138477
Change-Id: I930d378006970ed14bd61c9fd1c343c9b72bf8d2