Test apply power policy updates VHAL property
When power policy is applied, the VHAL property "CURRENT_POWER_POLICY"
should be updated to the new power policy ID. This test case ensures
this happens by verifying the daemon is notified that power policy is
changed.

Power policy daemon is also refactored to be a mock in testing instead
of a fake.

Bug: 300515753
Test: atest CarPowerManagementTest#testApplyPowerPolicy_vhalPropertyUpdated
Change-Id: I5198a8dd765cd50dd9fe5425b343ba9298b2a9c9