Add dynamic device parsing to audio zones helper
This parses the type and does some verification on device type (i.e.
is an dynamic output device or output bus). For legacy reasons, device
type of BUS still requires the device to exist at time of parsing. For
dynamic types the volume group will be constructed on the run but will
be assigned the actual audio device info when the configuration is
connected. Also more verification will be done on the default
configurations to have only bus devices, as the default configuration
must always be accessible for all zones.

Also modified check for primary zone configs, primary zone is
allowed to have multiple configurations as of version 4.

Bug: 305301155
Test: atest CarAudioZonesHelperTest
Change-Id: Ic7834cf9bbb2aa5fe0a250c51a3244ef77e9ac3b