Change routing audio policy construction logic
The logic now constructs the policy only for the currently selected
audio configuration and/or the default configuration for each zone.
Note that the default configuration is added so that it is always available
for routing even when the dynamic devices for the currently selected
configuration are not all active. The current rule that the default
audio configuration devices should be BUS type devices, which are always
available, allows for the default audio policy to remain as a
backup always.

In the next CLs, the audio configuration switching logic will be changed,
so that a new policy is constructed for the new configuration (new
switched configuration), prior to calling the switching logic to
completed. This new policy will replace the previous routing policy.

Bug: 305301155
Test: atest com.android.car.audio
Change-Id: Ibd329da04fc50d41430dd122c54144b2b1955631