Propagate audio device info to car audio zones
Used the audio device info update from audio service to update the audio
zones. In particular, the information is now propagated from the audio
zone, the car audio configuration, the volume groups, and finally to the
car audio device info.

Currently the logic is to update the audio device info as long as the
audio device type mathces. Since dynamic devices are not assigned audio
device address until they are connected there is no way to
distinguish otherwise. In the future this could be change if an specific
identifier is added. This assumption works well for BT devices, since
there can only be one BT device at a time.

One current assuption is that TYPE_BUS devices are always present, thus
bus type devices are not updated dynamically.

Bug: 305301155
Test: atest com.android.car.audio
Change-Id: I929bce4ab6a2871a0f7c4c52f185317af5386e2d