Create new audio policy on config change
Added logic to create new audio policy when an audio configuration
switch is updated. This allows for the audio policy with the updated
devices to be used. Also added the unregister of the previously
registered audio policy.

Next CLs will cover using the audio policy when a dynamic device
is removed. Given that the audio policy is constructed with the default
configuration, which should always be available, the switch should not
required re-creating the audio policy again.

Bug: 305301155
Test: atest com.android.car.audio
Change-Id: I38d2a3b4f3f027a2e1b1e00eec178781166b461b