Add repeated dynamic devices test to car audio service
While the dynamic device can be used on different car
configs for the same zone, they can not be used in the same
configuraiton for different volume groups. For this use case, it is
recommended to create a new volume group with all the required audio
usages for the dynamic device.

Bug: 305301155
Test: atest com.android.car.audio
Change-Id: Iede41dbce341ae494c24cf13daecced48148c67f