Fix car audio configuration with dynamic devices flag disabled
Fixed load car audio configuration with dynamic devices when dynamic
devices flag is disabled. This will load the car audio configurations
which do not have any dynamic devices if the dynamic
dynamic device flag is disabled.

Also fixed the car audio service unit tests which required the flag to
be disabled. The fix uses the set flag rule to disabled the flag before
the test is run.

Bug: 305301155
Test: atest com.android.car.audio
Change-Id: If26f761ce51eb334af7ebb7156697387f3fd9c32