Use ActivityOptions only on send
This reverts ag/26032749 (which intended to fix the
IllegalArgumentException on passing invalid options on create), and
instead removes the options on the creation completely as they have no
effect.
Since the code creates and sends the PendingIntent immediately from the
same process, the permissions of the sender and the receiver are the
same, so opt in by the sender is sufficient.

Test: atest android.car.app.cts.CarTaskViewControllerTest
Bug: 322986327
Change-Id: Iba168a5e8cc90d3fe867af54a78564f6110c81c4