Fix audio focus when OEM context are defined for car config
When OEM context and static audio context are not one-to-one the focus
interactions are lumped into the legacy context for the first usage in
the OEM context. This creates an issue since the interaction among
usages could them follow different rules depending on the OEM context
configuration. For example, if there was only a single media context
defined, then all the focus interaction among different usages would be
treated as media-media interaction. Refactor focus interactions to
use audio attribute usages instead of OEM car audio context, internally
the usage is converted to the appropiate legacy car audio context.

Bug: 323466501
Bug: 305301155
Test: atest CarAudioFocusTest com.android.car.audio
Change-Id: Ide279d7c9298cc087bb8d8b722a4b85cb28b58fd