Refactor fake daemon out of CPMS unit test
The FakeRefactoredCarPowerPolicyDaemon defined in the CPMS unit test suite can be used in other tests such as CarPowerManagementTest. Factoring out the class into its own file will allow it to be used in other tests.

Move TemporaryFile to car-test-lib to make it easier to use it in the
newly-define fake daemon.

Bug: 322558789
Test: atest CarPowerManagementServiceUnitTest
Change-Id: I9502e523452ad4d6a43f931c54a1274e3f2eef22