Fix CPM test for refactored policy logic
With the old logic, the CarPowerManagementTest was set up to use
MockedCarTestBase, which includes a fake power policy daemon. However,
due to the structure of the old logic, testing that's done on the Java
side will only test when CPMS has taken over from CPPD, so that fake
daemon doesn't have to do anything and the tests still pass. However,
this is different in the new logic, so the refactored daemon needs to be
used in MockedCarTestBase.

Bug: 322558789
Test: atest CarPowerManagementTest (refactoring flag enabled)
Change-Id: Ieae08171670d556113c278c4dadaae7a69d4cb83