Add preemptive policies header in policy reader dump
By removing the preemptive policies header in the PolicyReader dump for the new power policy logic, it messes with PowerPolicyHostTest's dump parser, which expects the preemptive power policy header to be there. By keeping the header, even when the refactored policy flag is enabled, it allows the dump parser to work correctly.

Bug: 316621864
Test: atest PowerPolicyHostTest
Change-Id: If7b7ec56d2082b7e2f0c73d44e0414d517571808