Remove callback when audio service is down
When the audio service is down some of the operations
(volume, focus, ducking, etc.) do not work as expected and could
cause an exception if performed from the car service thread. This change
will prevent callback externally to car audio service from performing some of
these operations. In particular, the change unregisters most of the system
callbacks to prevent the car audio service from performing some of these
operations while the audio server is down. For adio routing in particuar,
the car audio service first checks of the audio server indicator is in good
condition before the routing operation proceeds.

Bug: 326297826
Test: atest CarAudioServiceUnitTest
Change-Id: If19fbe1f9b9b7a99eaa07c2485b511801b25eba8