Clear cutout mode of Theme.DeviceDefault.Light.DarkActionBar
The windowLayoutInDisplayCutoutMode was set to default in
Theme.DeviceDefault.Light.DarkActionBar. That can make the app crash
since Android V. Because setting the cutout mode to non-always for a
non-floating activity window will get an IllegalArgumentException while
targeting V+.

So we can eidther set the attribute to always or leave it untouched.

Bug: 309578419
Bug: 327537204
Bug: 327538417
Bug: 327531666
Bug: 327669659
Bug: 327663836
Test: CTS
Change-Id: Ia30d695219884d0a0d98b5f0729cdc4785af2b50