Add car audio service check to car audio focus request
When the car zones audio focus is unregistered the audio policy and
service are set to null. This may lead to a race condition where the audio
focus request is being serviced (or on the looper message queue), and the
audio service/policy are null. Added a check to make sure the audio service
is not null before servicing the audio focus request. For the focus
request/abandon an error is logged but no other steps are taken. Note
that the audio policy is only null when not registered, thus even if we
wanted to reply there is no way to currently do so.

Bug: 326297826
Test: atest CarZonesAudioFocusUnitTest
Change-Id: I32b06dc6f2e141d32cc328d3e1fa501b20a726ec