Add flag to enable accessible volume in aaos
AAOS already separates the accessibility volume control from music
volume control. This allows for assistant and music to have separate
volumes control and audio routing. This changes explicitly enables flag
to allow separate accessibility volume control so that the audio system
and other apps are aware. The information can the be query from audio manager or the accessibility manager.

Bug: 333801000
Change-Id: I9e7c6e84156d24acc35d0d2f150954335fb447e5
Test: atest AudioManagerTest CarPackageManagerServiceTest
Test: atest CarPackageManagerTest