Audio: fix unmute by RAISE/UP
When using CoreVolumeGroup, while muted, any volume change from policy
volume callback (catching RAISE/UP) shall update the volume and unmute
if necessary. The mute state is updated and reported to HMI. However,
the unmute action is not performed. CoreVolumeGroup does not unmute
on setCurrentGainIndex (for the sake of orthogonality between volume and
mute) and expect explicit call to mute / unmute. applyMuteLocked is
called by mother class CarVolumeGroup. It is a no op for
CarAudioVolumeGroup, it mutes / unmutes for CoreVolumeGroup.

This CL fixes the issue by calling applyMuteLocked (in addition to
changing the mute state)  when a volume change is requested by VolumeGroup
mother class.

Bug: 334192996
Test atest CarVolumeGroupTest

Change-Id: I9bdccf244a653f4e020f54f629b64a186a9deb43
Signed-off-by: François Gaffie <francois.gaffie@renault.com>