Fix a race between two threads
When a HalCamera instance is trying to forward a newly captured frame
while a client is returning used frame buffers from multiple clients, it
is possible that two threads are racing to acquire the resouces held by
each other.

To avoid this scenario, this change modifies
VirtualCamera::doneWithFrame() not to call HalCamera::doneWithFrame()
while it holds the mutex; instead, it is delegated to a separate thread.

Bug: 346046198
Test: atest evsmanagerd_test --iteration=10
Flag: EXEMPT bugfix
Change-Id: Ide29260ae92f437e724c3d6d58631dc1596ce22a