Launch a registered camera activity as single task
StateMachine class requests launching a registered activity without
FLAG_ACTIVITY_MULTIPLE_TASK, and therefore prevents the system from
creating a new activity when a requested activity is already running.

This change is required to avoid situations that a running camera
activity is being demoted and frozen by ActivityManager after a user
profile is switched; to bring the activity back to the foreground after
a user switching procedure, CarEvsService listens to UserLifecycleEvent
and fires an intent to request a registered activity whenever a user
profile is unlocked while a backing event is still in progress.

testUserSwitchedWhileBackingEventInProgress is added to
CarEvsServiceUnitTest, to verify newly implemented behaviors.

Bug: 287511664
Test: 1. Create an additional user profile.
      2. Launch CarEvsCameraPreviwActivity with a system event; e.g. the
         backing event.
      3. Switching between user profiles with CarService command:
         > adb shell cmd car_service switch-user <user id>
Test: atest CarEvsServiceUnitTest
Flag: EXEMPT bugfix
Change-Id: I21f62c8c7a40bcad5d6ce7023162483eeebdecbc