Remove ensureTheCallerIsSignedWithPlatformKeys().
The method enforces that caller app must be signed with platform key.
We are changing the signing key of AA BRT from platform key to own key. So, it should be removed.

We reviewed and concluded this does not cause an additional seucrity issue. (Refer to the doc in the security review bug.)

Bug: 343846865 (Security review)
Bug: 343311831
Test: atest com.android.car.CarBugreportManagerServiceTest
Flag: EXEMPT minor behavioral change
Change-Id: I6d7b518ed4a87d52863fdba1479537980a90c708