Use separate thread for ConnectToVhal/getConfig.
Use a separate thread for connectToVhal and getPropConfig. getPropConfig
might be blocking for several seconds on initial boot on network or
on property discovery process. We use a separate thread for the blocking
operation to avoid blocking the main handler.

This CL also uses createVhalClient instead of tryCreate because we do
not rely on polling, but blocked on VHAL being ready in the separate
thread.

Flag: EXEMPT minor change that does not affect behavior
Test: atest WatchdogProcessServiceTest
Bug: 342470570
Change-Id: Ifae30bca317e30997219589ee26da197df21f368