Fix WatchdogPerfHandlerUnitTest flakiness in Cuttlefish
- To guarantee that all tests run after the WatchdogPerfHandlerUnitTest
setup method is complete, we run CarServiceUtils.runOnMainSync after
initializing the services. This ensures that the test thread remains
blocked until all tasks posted on the main thread are finished.

- WatchdogPerfHandlerUnitTest.crashAndDelayReconnectDaemon currently
uses postAtTime method which schedules a runnable to execute at a
specific absolute time passed as argument. However, due to the passage
of time since boot, the argument of 2000 milliseconds may have already
elapsed, leading to the immediate execution of the runnable. To ensure
that the runnable runs 2000 milliseconds in the future, we should use
postDelayed method instead.

- Refactored WatchdogPerfHandler.isConnectedToDaemon method such that
it doesn't unblock during interrupts or spurious wake-ups.

- In the current implementation of
WatchdogPerfHandler.onDaemonConnectionChange, the lock is acquired and
mIsConnectedToDaemon is set and then the lock is released. After this,
if isConnected is true we proceed to make a binder call. During this
period, if the thread waiting on isConnectedToDaemon() receives an
interrupt, it is unblocked and exits this method since
mIsConnectedToDaemon becomes true. This behavior is undesirable because
isConnectedToDaemon() should terminate when either lock.notifyAll() is
invoked or the timeout expires. To avoid this scenario, the entire
logic was synchronized in WatchdogPerfHandler.onDaemonConnectionChange.

Test: atest --rerun-until-failure 100 WatchdogPerfHandlerUnitTest
Bug: 337095515
Flag: TEST_ONLY
Change-Id: Iaa8a97eecbc8ae060aefb46a60992e2b057cd8f0