Prevent brightness change loop.
Add local buffers to store the brightness change we received from
VHAL and sent to VHAL in the recent time frame. If receive a change
from the other side (system or VHAL) that is equal to the stored
request, that means the change is caused by our request and we
should not issue a duplicate request again even though the current
value might be different value because a different request succeeded
first.

Flag: EXEMPT bug fix
Test: atest PowerHalServiceUnitTest
Bug: 358447023
Change-Id: Iba98ce08799a08593469a5bc19b1c2e70c91cc89