Fix crash in query statement for USB settings due to null bind value.
When the device supports AOAP mode, the serial is added to selectionArgs
but not checked in the condition. Previously, while building the query
in the queryFor method, an assumption was made: if the device supports
AOAP, a unique serial number exists. Our understanding is that this
assumption is incorrect. Hence, we verify for serial nullness first.

Bug: 341696108
Test: m
Flag: EXEMPT bugfix
Change-Id: I70639c91a10f8bfe35f52f699de7652d038bbc47