Delete Car Settings manifest entries related to super-obsolete zen onboarding
A notification was shown when upgrading from pre-P to P because default DND settings had changed. The default was later changed back so it was already obsolete at that point. Moreover, it should be impossible to upgrade from pre-P to V in one step.

Bug: 346789065
Test: N/A
Flag: EXEMPT Unflaggable -- removing resources, manifest entries, etc.
Change-Id: Ia709b9e1536439a408d21e1d919f860a9344260d