Always send initial value event for new subscription.
This CL fixes a previous undefined behavior where initial value event
may/may not be sent the client for a new subscription.

Previously, it is not guaranteed that the initial value event is
delivered to the first callback registered for [propId, areaId]. After
this CL, every new subscription is guaranteed to get the initial
value event.

Flag: android.car.feature.always_send_initial_value_event
Test: atest --host CarPropertyManagerUnitTest CarPropertyServiceUnitTest
Bug: 372530534
Change-Id: I548eaaa7464fa6178428ee5becc1187ab0801be6