Allow guest user creation regardless of no_add_user
CarUserService#createUser blocks callers with no_add_user from creating
a new user. This was causing an issue when a user with no_add_user tried
to remove itself, because the Setting UI first creates a guest user to
switch to, before it can delete the current user.
This change allows all users, regardless of the no_add_user restriction,
to create a guest user. This will not affect the current user experience
of guest user creation, because the quick settings UI to create a guest
user runs as the system user and bypasses the no_add_user restriction
check in CarUserService#createUser anyway.

Fixes: 319326658
Flag: EXEMPT bugfix

Test: manual verification using emulator:
      - Create a new user and switch back to Driver
      - Tap to "Profiles & accounts" in Settings
      - Tap to "Manage other profiles"
      - Select the users you created
      - Turn off the "Create new profiles" toggle button
      - Switch to the added user
      - Tap "Profiles & accounts" in Settings
      - Tap "Delete this profile"
Test: atest CarUserServiceTest
      atest android.car.cts.CarUserManagerTest

Change-Id: I1257c20415779f73c1632a60e4c45b4c67e6c2a0