Remove overlayed colors on Android Auto's keypad.
The new method for overlaying system colors was introduced with ag/29908603 but the olde overlay values do not work with this new styleless method.

Because original overlays were done to fix an issue with default color values that were incorrect in Android. Since this is fixed now, the overlays are removed for consistency. 

Bug: 284904653
Change-Id: I25233b2a232d7b4f1db391aaa25060ba7a1ecd48
Flag: EXEMPT bugfix
Test: Manual