Do not release real car service in CarServiceTest.
CarServiceTest is an isolated integration test that tests both
car-lib and car service. It is testing against a fake car service
instance running inside the test process and is not testing against
the real car service running in the system. As a result, it should
not interfere with the real system which might cause the whole
system to be unstable.

The original design is to stop the real car service so that the
test does not accidentally intefere with the real service however
it is not a good approach since the test now changes the system
state and is not a stand-alone test any more.

Flag: EXEMPT test change
Test: Presubmit
Bug: 380957733
Change-Id: Id0ffeb49caa9e8e4dc754d8c6d3d361002fd8817