Always sendConfigChangeEvent when it's user switch
In CarOccupantZoneService the first time call handleUserChangeLocked
the currentuser can be SystemUser 0 or DriverUser 11, later when user
switch from 0->11 happens, if previous set user is 11, there is no user
change, then the later logic sendConfigChangeEvent will not be called.

In this case, we always sendConfigChangeEvent when it's a user switch
regardless the return result from the handleUserChangeLocked.

Bug: 380898344
Flag: EXEMPT Bugfix
Test: manual test, run 'atest CarOccupantZoneServiceTest', and
      cts CarOccupantZoneManagerTest

Change-Id: Ib5d3ed07886724922683b35bf5fddb47ce078cff