Fix a race condition when binding to car service.
The issue is the clients facing DeadObjectException when interacting with car service, after car service crashed and restarted.
The issue is seen with both - clients that implement the onLifecycleChanged callback and clients that do not implement the callback.

Root cause:
When clients create the car object using Car.createCar API, they get `ready=true` before the car-library actually binds to car service.
Because of this, if car service dies while the bind operation is in progress, the binding fails with a `RemoteException`. But car-library does not get the `onServiceDisconnected` callback due to which clients face `DeadObjectException` when interacting with car service.

Resolution:
Force calling `onServiceDisconnected` for old car service connection, when onServiceConnected is received with the new car service binder.

Bug: 380289405

Change-Id: I037e50206e866d45cadd737f8fe6170c5aaa4d10
Signed-off-by: Alekhya Karavadi <akaravad@ford.com>