Use a system property to config cppd connect VHAL timeout.
Read a system property: carpowerpolicyd.connectvhal.Timeoutmillis
to config the timeout for car power policy daemon to connect to
VHAL. The default timeout is 5s and needs to be updated for slower
device (e.g. emulator).

In this CL, we set the default to 60s because we need to add the
config for emulator first before we change the default back to 5s.

Flag: EXEMPT simple change
Test: atest carpowerpolicyserver_test
Manual test on emulator.
Bug: 377348572

Change-Id: Ibfb8d5cec7375e134a6ad379811d90962c545fdf