Disable device manager when we disable default policy
Problem: The device manager is created even when the default policy is
disabled, which can lead to triggerConnections() calls due to the
broadcasts it signs up for and the way they're handled.

Solution: Only initialize the device manager when our policy is used. We
don't have any outward facing APIs for the device manager, and our
policy is the only one that calls it. All uses are null checked too.

Flag: EXEMPT, minor bug fix
Bug: 385390008
Test: atest com.android.car.bluetooth
Change-Id: Ia50575254be4d1e89f605113ef854ef7274a8e2f