Prevent returning car service binder to client.
Before car service binder is returned to the client, we need to
call linkToDeath to monitor its lifecycle. It is possible that the
binder is already dead before we call linkToDeath. In this case, we
should not return the dead binder to the client since the client
will never be notified about binder death (because it is already
dead). In this CL, we treat the dead binder as unuseful and ignore
it. If the call is waiting for a connection, we will keep waiting.

This CL also injects default create car timeout to reduce test time.

Test: atest CarUnitTest
Flag: EXEMPT bug fix
Bug: 398025989
Change-Id: I5815dbe760329c7154b0e5ab01897a0c992f0f51