Replace HandleThread with Executor
Replace vehicleHal handler thread with one single thread executor
and add a setCallbackExecutor method to allow the client to use
a different exeuctor. This is to allow PropertyHalService to use
a direct executor for property events in the future to avoid the
additional overhead introduced via handler thread.

This CL does not change the current behavior. All callbacks are
still invoked in sequential because we use one single thread
executor.

Note that this CL uses a local dispatch list instead of the
per-service dispatch list for property events because it is
possible that two handlePropertyEvents might be called at
the same time. Using a per-service list is not thread-safe
and error-prone. In reality, two onPropertyEvents will never
overlap due to how binder schedules one way methods. However,
it is possible that onPropertyEvents overlap with
injectVhalEvent.

Flag: EXEMPT refactor. This should cause no behavior change.
Test: atest VehicleHalTest
Bug: 402997840
Change-Id: Ic5aae86966fc90349ff2d736e9ae0bad1843fbaa