Assign default display power mode to always on.
Assign default display power mode to always on if not found from
settings. This is a temp fix for the bug where one of the passenger
display may be removed during early boot, causing its config to
be ignored (because we cannot find the display port). Then later,
when this display is added back, we will assign the default policy
to it.

We should change the default power mode back to on after this is
fixed.

Flag: EXEMPT bug fix
Test: atest ScreenOffHandlerUnitTest
Bug: 393626464
Change-Id: I0e809ab508b12bf98d27d2efb7689ac327c27e9f