Using cleaner to replace finalize.
Finalize is deprecated and error-prone. In fact, mock a Payload object
might cause finalize to be called on a LargeParcelableBase that is
not fully initialized (with a null mLock). This CL uses cleaner to
replace it.

This CL also updastes LargeParcelableTest to close the resouce.

This CL also adds 'implements interace' to LargeParcelableBase
subclass to provide clearer info.

Test: atest LargeParcelableTest
Flag: EXEMPT internal refactor
Bug: 402291632
Change-Id: I3efe6f52a2de5528242a8390d41687f316750602