Add a destroy method to HalServiceBase and CarSystemService.
We need a destroy method to act as an destructor. In production,
it will be called during onDestroy. In test, it should be called
by the test cases (usually tearDown). It is supposed to clean up
the resources allocated during constructor, for example, the
handler thread.

Flag: EXEMPT internal refactor.
Bug: 406550125
Test: atest VehicleHalTest ICarImplUnitTest
Change-Id: I815e568c59577531752ba53ec611235ce9ff2c68