Replace TimerThread with ScheduledThreadPoolExecutor.
There is no way to wait for the TimerThread to join so it is possible
that the timer thread is still handling property events after a test
case already finishes, which cause test flakiness. We should avoid
a handler thread from one test case from affecting another, so this
CL migrates the usage to ScheduledThreadPoolExecutor which we can
wait for termination when release is called.

Test: atest VehicleHalTest
Flag: EXEMPT test fix
Bug: 410430018
Change-Id: I702ab694efdb954a762492e3df7c601c7f4833c4