Add missing value of PowerManagerHelper.BRIGHTNESS_INVALID_FLOAT
Previously, `PowerManagerHelper.BRIGHTNESS_INVALID_FLOAT` had no value
specified in the signature file. That was due to a bug in Metalava
where it could not access the value of
`PowerManager.BRIGHTNESS_INVALID_FLOAT` which is `Float.NaN`. The
associated change in the topic fixes that issue which causes the value
to be added here.

The value looks a little strange but `(0.0f/0.0f)` is how `Float.NaN`
is defined.

Flag: EXEMPT bug fix not an API change
Bug: 354633349
Test: ./gradlew
Change-Id: I7e1d9a48355dad3de178d71ad85832b54462ba06