Get native package manager for every request
The reference to the native package manager that was being stored in the
init was becoming stale when testing in the 25Q2-swcar-dev branch.
The stale reference was causing EX_TRANSACTION_FAILED due to the dead
object, causing all init calls to fail.

Flag: EXEMPT does not introduce a new flag
Bug: 414613362
Test: manual
Change-Id: Id37eac025300ab3996a7e4b2c7c299d30b660133