CarWatchdog: Host settings string update logic in WatchdogPerfHandler.java.
When packages are disabled or enabled, the functionality to update
the settings string will be hosted in WatchdogPerfHandler.
This functionality will be invoked with IoOveruseHelper callback.

Rename processPackageChangedIntent method to onActionPackageChanged.
This makes the method name more readable and expresses the actual
intention.

Bug: 400460188
Test: atest WatchdogPerfHandlerUnitTest IoOveruseHandlerUnitTest
Flag: build.RELEASE_CAR_FRAMEWORK_WATCHDOG_FLASHMEMORY_REFACTOR
Change-Id: I068c67bb33b0e8254b4503b8e348df1f22732d53