Set config_allowPresentationOnInternalDisplay to false on AAOS
Now for the Connected Displays project, presentations are allowed
on internal displays, however, this could be a risk on some
platforms that have displays we never want to show presentations
on (e.g. a display showing the odometer on a car).

This change sets this behavior disabled for AAOS.

Flag: com.android.window.flags.enable_presentation_for_connected_displays
Bug: 429927513
Test: PresentationTest
Change-Id: I448b090ca971f49d5babe15a7ebde93fa3677e6d