Remove Bluetooth from auto PRODUCT_PACKAGES
Bluetooth is not exported to make as it is transitively installed via
`com.android.bt` apex, which is listed as PRODUCT_PACKAGES in
base_system.mk. In other words, listing Bluetooth in PRODUCT_PACKAGES is
a no-op because the module does not exist in make.

This change removes this invalid entry from PRODUCT_PACKAGES to make
packaging more closely aligned between soong only build and soong plus
make build.

Test: build/soong/scripts/soong_only_diff_test.py cf_x86_64_auto
Test: diff staging dir with/without the change
Bug: 421945305
Bug: 432083374
Flag: EXEMPT cleanup
Change-Id: Ibc4d490edd223008d1c608579c567926523054ea