Remove action bar from camera2 activities in KitchenSink
The action bar in CameraSystemActivity and MultiCameraPreviewActivity
was obscuring parts of the user interface, making it difficult to see
the full camera preview.

This change applies the KitchenSinkActivityTheme to these activities to
remove the action bar. This provides more screen space for the camera
previews, fixing the UI obstruction and improving the user experience
Test: tested manually
Bug:333609871
Flag: EXEMPT BUGFIX

Change-Id: I4b58232e4fb89501a4084a6e48c22f6d53b2edb5